tablet: fix build on macOS
This fixes the following error:

../../src/kudu/tablet/compaction-test.cc:1061:33: error: cannot initialize a parameter of type 'uint64_t *' (aka 'unsigned long long *') with an rvalue of type 'size_t *' (aka 'unsigned long *')
    ASSERT_OK(rs->CountLiveRows(&rs_live_rows));
                                ^~~~~~~~~~~~~

Reviewed-on: http://gerrit.cloudera.org:8080/16650
Tested-by: Attila Bukor <abukor@apache.org>

+++ src/kudu/tablet/compaction-test.cc
@@ -1055,9 +1055,9 @@ TEST_F(TestCompaction, TestMergeMRS) {
   unique_ptr<CompactionInput> input(CompactionInput::Merge(merge_inputs, &schema_));
   vector<shared_ptr<DiskRowSet>> result_rs;
   DoFlushAndReopen(input.get(), schema_, snap, kSmallRollThreshold, &result_rs);
-  int total_num_rows = 0;
+  uint64_t total_num_rows = 0;
   for (const auto& rs : result_rs) {
-    size_t rs_live_rows;
+    uint64_t rs_live_rows;
     ASSERT_OK(rs->CountLiveRows(&rs_live_rows));
     total_num_rows += rs_live_rows;
   }
